Use unix timestamp instead of just HHmmss (part 2)

+++ common/connection_manager.cpp
@@ -9,8 +9,8 @@ ConnectionManager::ConnectionManager(const QList<QObject *> &inputs, const QList
             senders << input;
             for (auto * strategy : strategies) {
                 if (strategy != nullptr) {
-                    QObject::connect(input, SIGNAL(newMarketData(QString,int,double,int,double,int,double,int)),
-                                     strategy, SLOT(onMarketData(QString,int,double,int,double,int,double,int)));
+                    QObject::connect(input, SIGNAL(newMarketData(QString,qint64,double,int,double,int,double,int)),
+                                     strategy, SLOT(onMarketData(QString,qint64,double,int,double,int,double,int)));
                     QObject::connect(input, SIGNAL(tradingDayChanged(QString)),
                                      strategy, SLOT(setTradingDay(QString)));
                 }
