+++ future_arbitrageur/depth_market.h
@@ -5,7 +5,7 @@
 
 class DepthMarket {
 public:
-    int   time;
+    qint64 time;
     double lastPrice;
     double askPrice;
     int    askVolume;
@@ -13,7 +13,7 @@ public:
     int    bidVolume;
 
     DepthMarket();
-    DepthMarket(int time, double lastPrice, double askPrice, int askVolume, double bidPrice, int bidVolume);
+    DepthMarket(qint64 time, double lastPrice, double askPrice, int askVolume, double bidPrice, int bidVolume);
     bool significantChange(const DepthMarket &other) const;
     bool isUpperLimit() const { return bidVolume == 0; }
     bool isLowerLimit() const { return askVolume == 0; }
